Remove thresholds from configs entirely, refactor types
