+++ src/kayenta/domain/ICanaryScoreThresholds.ts
@@ -1,0 +1,4 @@
+export interface ICanaryScoreThresholds {
+  pass: number;
+  marginal: number;
+}
