+++ src/kayenta/domain/index.ts
@@ -1,13 +1,15 @@
 export * from './ICanaryConfig';
+export * from './ICanaryScoreThresholds';
 export * from './ICanaryConfigSummary';
 export * from './ICanaryJudgeResult';
 export * from './ICanaryJudgeResultSummary';
 export * from './IJudge';
 export * from './IKayentaAccount';
 export * from './IMetricSetPair';
-export * from './ICanaryJudgeStage';
 export * from './ISetupCanaryStage';
 export * from './ICanaryExecutionStatusResult';
 export * from './IMetricsServiceMetadata';
 export * from './ICanaryConfigUpdateResponse';
 export * from './IKayentaStageConfig';
+export * from './MetricClassificationLabel';
+export * from './ScoreClassificationLabel';
